requirements.txt
pandas
pyperclip
pyautogui
Pillow
pywin32
